backend1/requirements.txt
fastapi
uvicorn
tensorflow==2.12
Pillow
numpy
keras
python-multipart
